requirements.txt
openai==1.62.0
google-genai==1.2.0
mistralai==1.5.0
python-dotenv==1.0.1
